refectoring get repository
